Small rearrangement of the functions
