+++ pycode/memilio-surrogatemodel/memilio/surrogatemodel/ode_secir_simple/model.py
@@ -26,8 +26,175 @@ import pandas as pd
 import tensorflow as tf
 
 from memilio.simulation.osecir import InfectionState
-from memilio.surrogatemodel.ode_secir_simple import network_architectures
+import memilio.surrogatemodel.ode_secir_simple.network_architectures as architectures 
 
+def initialize_model(parameters): 
+    """ Initialize model from given list of parameters
+
+    :param parameters: tuple of parameters describing the model architecture, it should be of the form 
+            (num_days_per_output, num_outputs, num_hidden_layers, neurons_per_layer, name_activation, name_architecture)
+    :returns: tensor flow keras model with the given architecture 
+     """
+    label_width, num_outputs, layer, neuron_number, activation, modelname = parameters
+    if modelname == "Dense":
+        return architectures.mlp_multi_input_multi_output(
+            label_width = label_width, 
+            num_outputs = num_outputs,
+            num_hidden_layers = layer, 
+            num_neurons_per_layer = neuron_number,
+            activation = activation
+        )
+    elif modelname == "LSTM":
+        return architectures.lstm_multi_input_multi_output(
+            label_width = label_width, 
+            num_outputs = num_outputs,
+            num_hidden_layers = layer, 
+            num_neurons_per_layer = neuron_number,
+            activation = activation
+        )
+    elif modelname == "CNN":
+        return architectures.cnn_multi_input_multi_output(
+            label_width = label_width, 
+            num_outputs = num_outputs,
+            num_hidden_layers = layer, 
+            num_neurons_per_layer = neuron_number,
+            activation = activation
+        )
+
+def network_fit(model, inputs, labels, training_parameter, plot=True):
+    """ Training and evaluation of a given model with mean squared error loss and Adam optimizer using the mean absolute error as a metric.
+
+    :param model: Keras sequential model.
+    :param inputs: Training input data 
+    :param labels: Training label data 
+    :param training_parameter: tuple of parameters used for the training process, it should be of the form
+        (early_stop, max_epochs, loss, optimizer, metrics), where loss is a loss-function implemented in keras, optimizer is the name of the used optimizer, 
+        metrics is a list of used training metrics, e.g. [tf.keras.metrics.MeanAbsoluteError(), tf.keras.metrics.MeanAbsolutePercentageError()]
+    :param plot:  (Default value = True)
+
+    """
+    early_stop, max_epochs, loss, optimizer, metrics = training_parameter
+    data_splitted = split_data(inputs, labels)
+
+    train_inputs = data_splitted['train_inputs']
+    train_labels = data_splitted['train_labels']
+    valid_inputs = data_splitted['valid_inputs']
+    valid_labels = data_splitted['valid_labels']
+    test_inputs = data_splitted['test_inputs']
+    test_labels = data_splitted['test_labels']
+
+    early_stopping = tf.keras.callbacks.EarlyStopping(monitor='val_loss',
+                                                      patience=early_stop,
+                                                      mode='min')
+
+    model.compile(
+        loss=loss,
+        optimizer=optimizer,
+        metrics=metrics)
+
+    history = model.fit(train_inputs, train_labels, epochs=max_epochs,
+                        validation_data=(valid_inputs, valid_labels),
+                        callbacks=[early_stopping])
+
+    if (plot):
+        plot_losses(history)
+        plot_compartment_prediction_model(
+            test_inputs, test_labels, model=model,
+            plot_compartment='InfectedSymptoms', max_subplots=3)
+        df = get_test_statistic(test_inputs, test_labels, model)
+        print(df)
+
+    return history
+
+
+def network_fit1(path, model, max_epochs=30, early_stop=500, plot=True):
+    """ Training and evaluation of a given model with mean squared error loss and Adam optimizer using the mean absolute error as a metric.
+
+    :param path: path of the dataset.
+    :param model: Keras sequential model.
+    :param max_epochs: int maximum number of epochs in training. (Default value = 30)
+    :param early_stop: Integer that forces an early stop of training if the given number of epochs does not give a significant reduction of validation loss. (Default value = 500)
+    :param plot:  (Default value = True)
+
+    """
+
+    if not os.path.isfile(os.path.join(path, 'data_secir_simple.pickle')):
+        ValueError("no dataset found in path: " + path)
+
+    file = open(os.path.join(path, 'data_secir_simple.pickle'), 'rb')
+
+    data = pickle.load(file)
+    data_splitted = split_data(data['inputs'], data['labels'])
+
+    train_inputs = data_splitted['train_inputs']
+    train_labels = data_splitted['train_labels']
+    valid_inputs = data_splitted['valid_inputs']
+    valid_labels = data_splitted['valid_labels']
+    test_inputs = data_splitted['test_inputs']
+    test_labels = data_splitted['test_labels']
+
+    early_stopping = tf.keras.callbacks.EarlyStopping(monitor='val_loss',
+                                                      patience=early_stop,
+                                                      mode='min')
+
+    model.compile(
+        loss=tf.keras.losses.MeanSquaredError(),
+        optimizer=tf.keras.optimizers.Adam(),
+        metrics=[tf.keras.metrics.MeanAbsoluteError()])
+
+    history = model.fit(train_inputs, train_labels, epochs=max_epochs,
+                        validation_data=(valid_inputs, valid_labels),
+                        callbacks=[early_stopping])
+
+    if (plot):
+        plot_losses(history)
+        plot_compartment_prediction_model(
+            test_inputs, test_labels, model=model,
+            plot_compartment='InfectedSymptoms', max_subplots=3)
+        df = get_test_statistic(test_inputs, test_labels, model)
+        print(df)
+    return history
+
+def split_data(inputs, labels, split_train=0.7,
+               split_valid=0.2, split_test=0.1):
+    """ Split data set in training, validation and testing data sets.
+
+    :param inputs: input dataset
+    :param labels: label dataset
+    :param split_train: Share of training data sets. (Default value = 0.7)
+    :param split_valid: Share of validation data sets. (Default value = 0.2)
+    :param split_test: Share of testing data sets. (Default value = 0.1)
+    :returns: dictionary of the form {'train_inputs', 'train_labels', 'valid_inputs',
+        'valid_labels', 'test_inputs', 'test_labels'}
+    """
+
+    if split_train + split_valid + split_test > 1 + 1e-10:
+        raise ValueError(
+            "Summed data set shares are greater than 1. Please adjust the values.")
+    elif inputs.shape[0] != labels.shape[0] or inputs.shape[2] != labels.shape[2]:
+        raise ValueError(
+            "Number of batches or features different for input and labels")
+
+    n = inputs.shape[0]
+    n_train = int(n * split_train)
+    n_valid = int(n * split_valid)
+    n_test = n - n_train - n_valid
+
+    inputs_train, inputs_valid, inputs_test = tf.split(
+        inputs, [n_train, n_valid, n_test], 0)
+    labels_train, labels_valid, labels_test = tf.split(
+        labels, [n_train, n_valid, n_test], 0)
+
+    data = {
+        'train_inputs': inputs_train,
+        'train_labels': labels_train,
+        'valid_inputs': inputs_valid,
+        'valid_labels': labels_valid,
+        'test_inputs': inputs_test,
+        'test_labels': labels_test
+    }
+
+    return data
 
 def plot_compartment_prediction_model(
         inputs, labels, model=None, plot_compartment='InfectedSymptoms',
@@ -87,56 +254,6 @@ def plot_compartment_prediction_model(
         os.mkdir("plots")
     plt.savefig('plots/evaluation_secir_simple_' + plot_compartment + '.png')
 
-
-def network_fit(path, model, max_epochs=30, early_stop=500, plot=True):
-    """ Training and evaluation of a given model with mean squared error loss and Adam optimizer using the mean absolute error as a metric.
-
-    :param path: path of the dataset.
-    :param model: Keras sequential model.
-    :param max_epochs: int maximum number of epochs in training. (Default value = 30)
-    :param early_stop: Integer that forces an early stop of training if the given number of epochs does not give a significant reduction of validation loss. (Default value = 500)
-    :param plot:  (Default value = True)
-
-    """
-
-    if not os.path.isfile(os.path.join(path, 'data_secir_simple.pickle')):
-        ValueError("no dataset found in path: " + path)
-
-    file = open(os.path.join(path, 'data_secir_simple.pickle'), 'rb')
-
-    data = pickle.load(file)
-    data_splitted = split_data(data['inputs'], data['labels'])
-
-    train_inputs = data_splitted['train_inputs']
-    train_labels = data_splitted['train_labels']
-    valid_inputs = data_splitted['valid_inputs']
-    valid_labels = data_splitted['valid_labels']
-    test_inputs = data_splitted['test_inputs']
-    test_labels = data_splitted['test_labels']
-
-    early_stopping = tf.keras.callbacks.EarlyStopping(monitor='val_loss',
-                                                      patience=early_stop,
-                                                      mode='min')
-
-    model.compile(
-        loss=tf.keras.losses.MeanSquaredError(),
-        optimizer=tf.keras.optimizers.Adam(),
-        metrics=[tf.keras.metrics.MeanAbsoluteError()])
-
-    history = model.fit(train_inputs, train_labels, epochs=max_epochs,
-                        validation_data=(valid_inputs, valid_labels),
-                        callbacks=[early_stopping])
-
-    if (plot):
-        plot_losses(history)
-        plot_compartment_prediction_model(
-            test_inputs, test_labels, model=model,
-            plot_compartment='InfectedSymptoms', max_subplots=3)
-        df = get_test_statistic(test_inputs, test_labels, model)
-        print(df)
-    return history
-
-
 def plot_losses(history):
     """ Plots the losses of the model training.
 
@@ -182,61 +299,45 @@ def get_test_statistic(test_inputs, test_labels, model):
     return mean_percentage
 
 
-def split_data(inputs, labels, split_train=0.7,
-               split_valid=0.2, split_test=0.1):
-    """ Split data set in training, validation and testing data sets.
-
-    :param inputs: input dataset
-    :param labels: label dataset
-    :param split_train: Share of training data sets. (Default value = 0.7)
-    :param split_valid: Share of validation data sets. (Default value = 0.2)
-    :param split_test: Share of testing data sets. (Default value = 0.1)
-
-    """
-
-    if split_train + split_valid + split_test > 1 + 1e-10:
-        raise ValueError(
-            "Summed data set shares are greater than 1. Please adjust the values.")
-    elif inputs.shape[0] != labels.shape[0] or inputs.shape[2] != labels.shape[2]:
-        raise ValueError(
-            "Number of batches or features different for input and labels")
-
-    n = inputs.shape[0]
-    n_train = int(n * split_train)
-    n_valid = int(n * split_valid)
-    n_test = n - n_train - n_valid
-
-    inputs_train, inputs_valid, inputs_test = tf.split(
-        inputs, [n_train, n_valid, n_test], 0)
-    labels_train, labels_valid, labels_test = tf.split(
-        labels, [n_train, n_valid, n_test], 0)
-
-    data = {
-        'train_inputs': inputs_train,
-        'train_labels': labels_train,
-        'valid_inputs': inputs_valid,
-        'valid_labels': labels_valid,
-        'test_inputs': inputs_test,
-        'test_labels': labels_test
-    }
-
-    return data
-
-
 if __name__ == "__main__":
+    label_width = 30 
+    num_outputs = 8
+    # General grid search parameters for the training process:
+    early_stop = [100]
+    max_epochs = [3]
+    losses=[tf.keras.losses.MeanAbsolutePercentageError()]
+    optimizers = ['AdamW']
+    metrics=[[tf.keras.metrics.MeanAbsoluteError(), tf.keras.metrics.MeanAbsolutePercentageError()]]
+
+    # Define grid search parameters for the architecture
+    hidden_layers = [3]
+    neurons_in_hidden_layer = [32]
+    activation_function = ['relu']
+    models = ["LSTM"]
+
+    # Collecting parameters
+    training_parameters = [(early, epochs, loss, optimizer, metric) 
+                           for early in early_stop for epochs in max_epochs for loss in losses 
+                           for optimizer in optimizers for metric in metrics]
+
+    model_parameters = [(label_width, num_outputs, layer, neuron_number, activation, modelname)
+                        for layer in hidden_layers for neuron_number in neurons_in_hidden_layer 
+                        for activation in activation_function for modelname in models]
+
+    # getting training data
+    filename = "data_secir_simple_30days_10k.pickle"
     path = os.path.dirname(os.path.realpath(__file__))
     path_data = os.path.join(os.path.dirname(os.path.realpath(
         os.path.dirname(os.path.realpath(path)))), 'data')
-    max_epochs = 400
 
-    model = "LSTM"
-    if model == "Dense":
-        model = network_architectures.mlp_multi_input_single_output()
-    elif model == "LSTM":
-        model = network_architectures.lstm_multi_input_multi_output(30)
-    elif model == "CNN":
-        model = network_architectures.cnn_multi_input_multi_output(30)
+    if not os.path.isfile(os.path.join(path_data, filename)):
+        raise ValueError(f"No dataset found in path: {path_data}")
 
-    model_output = network_fit(
-        path_data, model=model,
-        max_epochs=max_epochs)
+    with open(os.path.join(path_data, filename), 'rb') as file:
+        data = pickle.load(file)
+
+    input_data = data['inputs']
+    label_data = data['labels']
+    model = initialize_model(model_parameters[0])
+    model_output = network_fit(model, input_data, label_data, training_parameters[0], False)
+    plot_compartment_prediction_model(input_data, label_data, model)
